% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{il_edges}
\alias{il_edges}
\title{Deaths in the Iliad 3}
\format{
A tbl_graph object.

In this version of the data, arcs are undirected ties between killer and killed
}
\source{
Gabriel Rossman, "Glory and Gore", *Contexts* October 2017 https://journals.sagepub.com/doi/full/10.1177/1536504217732052
}
\usage{
il_edges
}
\description{
Deaths in the Iliad: undirected
}
\keyword{datasets}
