% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{il_killed_by}
\alias{il_killed_by}
\title{Deaths in the Iliad 2}
\format{A tbl_graph object.
\describe{
In this version of the data, arcs signify loser -> winner
}}
\source{
Gabriel Rossman, "Glory and Gore", *Contexts* October 2017 https://journals.sagepub.com/doi/full/10.1177/1536504217732052
}
\usage{
il_killed_by
}
\description{
Deaths in the Iliad: loser -> winner, and all labels
}
\keyword{datasets}
