% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{il_tidy}
\alias{il_tidy}
\title{Deaths in the Iliad 4}
\format{
A tbl_graph object.

Tidygraph version of the Iliad data
}
\source{
Gabriel Rossman, "Glory and Gore", *Contexts* October 2017 https://journals.sagepub.com/doi/full/10.1177/1536504217732052
}
\usage{
il_tidy
}
\description{
Deaths in the Iliad: tidy representation
}
\keyword{datasets}
