% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{revere_groups}
\alias{revere_groups}
\title{Pre-Revolutionary Boston Clubs: group x group}
\format{A tbl_graph object.}
\source{
David Hackett Fischer, Paul Reviere's Ride (OUP 1995)
}
\usage{
revere_groups
}
\description{
Membership or affiliation in Pre-Revolutionary organizations,
from Appendix D of David Hackett Fischer's "Paul Revere's Ride"
(Oxford University Press, 1995)
}
\keyword{datasets}
