% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{socjobs}
\alias{socjobs}
\title{Sociology Top 25 Assistant Professor Job Placements 1990-2017}
\format{A tibble with 343 rows and 7 columns
\describe{
\item{\code{sex}}{Sex of individual.}
\item{\code{job_yr}}{Year of first Assistant Professor position.}
\item{\code{job_dept}}{Hiring department}
\item{\code{phd_yr}}{Year of Ph.D.}
\item{\code{phd_dept}}{Individual's Ph.D department}
\item{\code{top25phd}}{Is PhD Department a Top 25 USNWR Soc program?}
\item{\code{gap}}{Years elapsed between PhD and TT Job.}
}}
\source{
John Robert Warren, (2019) "How Much Do You Have to Publish to Get a Job in a Top Sociology Department? Or to Get Tenure? Trends Over a Generation", *Sociological Science* February 27, 2019 doi://10.15195/v6.a7. Original data files available at <https://www.rob-warren.com/pub_trends.html>.
}
\usage{
socjobs
}
\description{
Limited Placement data for Sociology (TT AP Jobs in USNWR Top 25 departments only)
}
\keyword{datasets}
